refactor(baseballnumber): BaseballNumbers 객체 생성 로직 분리
1. Computer, Input에 있던 BaseballNumbers 생성 로직을 BaseballNumbersGenerator 객체로 분리
