+++ src/test/java/baseball/InputTest.java
@@ -9,22 +9,12 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class InputTest {
-	@Test
-	void convert_string_to_numeric_list() {
-		input("123");
-
-		BaseballNumbers baseballNumbers = Input.readPlayerBaseballNumbers();
-
-		assertThat(baseballNumbers.getBaseballNumbers())
-			.containsExactly(1, 2, 3);
-	}
-
 	@DisplayName("중복된 값이 입력되면 예외가 발생한다")
 	@Test
 	void given_duplicated_numbers_then_throw_exception() {
 		input("122");
 		assertThatExceptionOfType(IllegalArgumentException.class)
-			.isThrownBy(() -> Input.readPlayerBaseballNumbers())
+			.isThrownBy(Input::readPlayerBaseballNumbers)
 			.withMessage("입력 된 숫자에 중복이 있으면 안됩니다.");
 	}
 
@@ -33,7 +23,7 @@ class InputTest {
 	void given_number_other_than_three_then_throw_exception() {
 		input("12d");
 		assertThatExceptionOfType(IllegalArgumentException.class)
-			.isThrownBy(() -> Input.readPlayerBaseballNumbers())
+			.isThrownBy(Input::readPlayerBaseballNumbers)
 			.withMessage("3개의 숫자(1~9)를 입력 하세요.");
 	}
 
@@ -42,7 +32,7 @@ class InputTest {
 	void given_not_between_1_and_9_then_throw_exception() {
 		input("109");
 		assertThatExceptionOfType(IllegalArgumentException.class)
-			.isThrownBy(() -> Input.readPlayerBaseballNumbers())
+			.isThrownBy(Input::readPlayerBaseballNumbers)
 			.withMessage("3개의 숫자(1~9)를 입력 하세요.");
 	}
 
@@ -51,7 +41,7 @@ class InputTest {
 	void when_no_input_value_then_throw_exception() {
 		input("");
 		assertThatExceptionOfType(NoSuchElementException.class)
-			.isThrownBy(() -> Input.readPlayerBaseballNumbers());
+			.isThrownBy(Input::readPlayerBaseballNumbers);
 	}
 
 	private void input(String value) {
